generalize port range
