+++ localstack/utils/net.py
@@ -1,13 +1,15 @@
 import logging
 import re
 import socket
+import threading
 from contextlib import closing
 from typing import List, Optional, Union
 from urllib.parse import urlparse
 
+import cachetools
 import dns.resolver
 
-from localstack.utils.generic.wait_utils import retry
+from .sync import retry
 
 LOG = logging.getLogger(__name__)
 
@@ -174,3 +176,59 @@ def is_ipv4_address(address: str) -> bool:
     :return: True if string looks like IPv4 address, False otherwise
     """
     return bool(re.match(IP_REGEX, address))
+
+
+class PortNotAvailableException(Exception):
+    """Exception which indicates that the PortPool could not reserve a port."""
+
+    pass
+
+
+class PortRange:
+    """Manages a range of ports that can be reserved and requested."""
+
+    def __init__(self, start: int, end: int):
+        # cache for locally available ports (ports are reserved for a short period of a few seconds)
+        self._PORTS_CACHE = cachetools.TTLCache(maxsize=100, ttl=6)
+        self._PORTS_LOCK = threading.RLock()
+        self.start = start
+        self.end = end
+
+    def reserve_port(self, port: int = None) -> int:
+        """
+        Reserves the given port (if it is still free). If the given port is None, it reserves a free port from the
+        configured port range for external services. If a port is given, it has to be within the configured
+        range of external services (i.e., in the range [self.start, self.end)).
+        :param port: explicit port to check or None if a random port from the configured range should be selected
+        :return: reserved, free port number (int)
+        :raises: PortNotAvailableException if the given port is outside the configured range, it is already bound or
+                    reserved, or if the given port is none and there is no free port in the configured service range.
+        """
+        ports_range = range(self.start, self.end)
+        if port is not None and port not in ports_range:
+            raise PortNotAvailableException(
+                f"The requested port ({port}) is not in the port range ({ports_range})."
+            )
+        with self._PORTS_LOCK:
+            if port is not None:
+                return self._check_port(port)
+            else:
+                for port_in_range in ports_range:
+                    try:
+                        return self._check_port(port_in_range)
+                    except PortNotAvailableException:
+                        # We ignore the fact that this single port is reserved, we just check the next one
+                        pass
+        raise PortNotAvailableException(
+            "No free network ports available in the port range (currently reserved: %s)",
+            list(self._PORTS_CACHE.keys()),
+        )
+
+    def _check_port(self, port: int) -> int:
+        """Checks if the given port is currently not reserved and can be bound."""
+        if not self._PORTS_CACHE.get(port) and port_can_be_bound(port):
+            # reserve the port for a short period of time
+            self._PORTS_CACHE[port] = "__reserved__"
+            return port
+        else:
+            raise PortNotAvailableException(f"The given port ({port}) is already reserved.")
